+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.0.4
+  * `EnvyProperty.updateValues` no longer includes unavailable data when finishing
+  * Bumped `polymer` dependency to `^1.0.0-rc.4`
+  * Added `NakedProperty`, an `EnvyProperty` that makes its private values available 
+
 ### 0.0.3
   * Removed null operator ambiguity